Fixed all the bugs! And introduced some new ones.
Units can no longer walk through walls! But they bounce and slide around
a lot, and it's impossible to figure out what's going on unless they're
more visually distinct, so we should do that, and it would be better if
they collided like spheres or capsules or something else round.